new input function

+++ src/FPSTimer.h
@@ -8,16 +8,16 @@
 class FPSTimer : public Timer
 {
 private:
-    Uint32 previous_time = 0;
-    const Uint32 time_for_frame =  1000 / MAX_FPS;
-    Uint32 delta = 0;
+    Uint64 previous_time = 0;
+    const Uint64 time_for_frame =  1000 / MAX_FPS;
+    Uint64 delta = 0;
 public:
 
     FPSTimer()= default;
 
     float get_delta() const;
 
-    void clamp_and_print_fps(const int& amount_of_frames);
+    void clamp_and_print_fps(const long int& amount_of_frames);
 
 };
